requirements.txt
pywin32>=306
Pillow>=10.0.0
pyinstaller>=6.0.0
xlsxwriter>=3.1.0
fpdf==1.7.2
customtkinter>=5.2.0
llama-cpp-python>=0.3.0
langchain>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.4.0
langchain-classic>=1.0.0
pydantic>=2.0.0
instructor>=1.0.0
